requirements.txt
asgiref==3.3.1
Django==3.1.6
django-environ==0.4.5
django-filter==2.4.0
gunicorn==20.0.4
install==1.3.4
numpy==1.20.1
pandas==1.2.2
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
